+++ src/main/kotlin/com/github/oxc/project/oxcintellijplugin/oxfmt/lsp/OxfmtLspServerSupportProvider.kt
@@ -1,19 +1,17 @@
 package com.github.oxc.project.oxcintellijplugin.oxfmt.lsp
 
+import com.github.oxc.project.oxcintellijplugin.OxcLspServerPool
+import com.github.oxc.project.oxcintellijplugin.OxcLspTool
 import com.github.oxc.project.oxcintellijplugin.OxcIcons
 import com.github.oxc.project.oxcintellijplugin.oxfmt.OxfmtPackage
 import com.github.oxc.project.oxcintellijplugin.oxfmt.settings.OxfmtConfigurable
 import com.github.oxc.project.oxcintellijplugin.oxfmt.settings.OxfmtSettings
-import com.intellij.javascript.nodejs.library.yarn.pnp.YarnPnpNodePackage
 import com.intellij.openapi.diagnostic.thisLogger
 import com.intellij.openapi.project.Project
-import com.intellij.openapi.roots.ProjectRootManager
 import com.intellij.openapi.vfs.VirtualFile
-import com.intellij.openapi.vfs.VirtualFileManager
 import com.intellij.platform.lsp.api.LspServer
 import com.intellij.platform.lsp.api.LspServerSupportProvider
 import com.intellij.platform.lsp.api.lsWidget.LspServerWidgetItem
-import kotlin.io.path.Path
 
 class OxfmtLspServerSupportProvider : LspServerSupportProvider {
 
@@ -30,22 +28,7 @@ class OxfmtLspServerSupportProvider : LspServerSupportProvider {
         if (!oxfmt.isEnabled()) {
             return
         }
-        val executable = oxfmt.binaryPath(file) ?: return
-        val nodePackage = oxfmt.getPackage(file)
-        val root = if (nodePackage != null) {
-            if (nodePackage is YarnPnpNodePackage) {
-                nodePackage.getPackageJson(project)?.parent ?: return
-            } else {
-                VirtualFileManager.getInstance()
-                    .findFileByNioPath(Path(nodePackage.systemIndependentPath))?.parent?.parent
-                ?: return
-            }
-        } else {
-            ProjectRootManager.getInstance(project).fileIndex.getContentRootForFile(file) ?: return
-        }
-
-        serverStarter.ensureServerStarted(
-            OxfmtLspServerDescriptor(project, root, executable, oxfmt.binaryParameters(file)))
+        OxcLspServerPool.getInstance(project).fileOpened(OxcLspTool.OXFMT, file, allowEviction = false)
     }
 
     override fun createLspServerWidgetItem(lspServer: LspServer,